feat(README): Add readme to project

+++ Models/Livre.cs
@@ -7,7 +7,7 @@
         public string  Titre {get; set;}
 
         public string Auteur { get; set; }
-        public required int Isbn { get; set; }
+        public required long Isbn { get; set; }
 
         public bool Disponibilite { get; set; }
     }
